+++ src/parser/parsing/line.go
@@ -70,70 +70,3 @@ func splitOffPrecedingWhitespace(line string) (string, string) {
 	})
 	return text, line[:len(line)-len(text)]
 }
-
-func Join(ls []Line) string {
-	result := ""
-	for _, l := range ls {
-		result += l.Original()
-	}
-	return result
-}
-
-func IsBlank(l Line) bool {
-	if len(l.Text) == 0 {
-		return true
-	}
-	for _, c := range l.Text {
-		if c != ' ' && c != '\t' {
-			return false
-		}
-	}
-	return true
-}
-
-func Insert(ls []Line, position int, text string, isIndented bool, prefs Preferences) []Line {
-	if position > len(ls)+1 {
-		panic("Out of bounds")
-	}
-	result := make([]Line, len(ls)+1)
-	offset := 0
-	for i := range result {
-		if i == position {
-			line := ""
-			if isIndented {
-				line += prefs.Indentation
-			}
-			line += text + prefs.LineEnding
-			result[i] = NewLineFromString(line, i+1)
-			offset = 1
-		} else {
-			result[i] = ls[i-offset]
-		}
-		result[i].LineNumber = i + 1
-	}
-	if position > 0 && result[position-1].originalLineEnding == "" {
-		result[position-1].originalLineEnding = prefs.LineEnding
-	}
-	return result
-}
-
-type Preferences struct {
-	LineEnding  string
-	Indentation string
-}
-
-func DefaultPreferences() Preferences {
-	return Preferences{
-		LineEnding:  "\n",
-		Indentation: "    ",
-	}
-}
-
-func (p *Preferences) Adapt(l *Line) {
-	if l.IndentationLevel() == 1 {
-		p.Indentation = l.originalIndentation
-	}
-	if l.originalLineEnding != "" {
-		p.LineEnding = l.originalLineEnding
-	}
-}
